Split EnumValueRefParser out of ValueRefParser
This should reduce the memory load on compilers.
