+++ msvc2015/Parsers/Parsers.vcxproj.filters
@@ -128,6 +128,9 @@
     <ClCompile Include="..\..\parse\ValueRefParser.cpp">
       <Filter>Source Files\parse</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\parse\EnumValueRefRules.cpp">
+      <Filter>Source Files\parse</Filter>
+    </ClCompile>
     <ClCompile Include="..\..\parse\EmpireStatsParser.cpp">
       <Filter>Source Files\parse</Filter>
     </ClCompile>
@@ -227,6 +230,9 @@
       <Filter>Header Files\parse</Filter>
     </ClInclude>
     <ClInclude Include="..\..\parse\ValueRefParser.h">
+      <Filter>Header Files\parse</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\parse\EnumValueRefRules.h">
       <Filter>Header Files\parse</Filter>
     </ClInclude>
     <ClInclude Include="..\..\parse\EffectParserImpl.h">
