+++ parse/ConditionParser6.h
@@ -2,6 +2,7 @@
 #define _ConditionParser6_h_
 
 #include "ConditionParserImpl.h"
+#include "EnumValueRefRules.h"
 
 namespace parse { namespace detail {
     struct condition_parser_rules_6 : public condition_parser_grammar {
