+++ parse/EffectParser3.h
@@ -2,6 +2,7 @@
 #define _EffectParser3_h_
 
 #include "EffectParserImpl.h"
+#include "EnumValueRefRules.h"
 
 namespace parse { namespace detail {
     struct effect_parser_rules_3 : public effect_parser_grammar {
