+++ parse/SpeciesParser.cpp
@@ -3,6 +3,7 @@
 #include "ParseImpl.h"
 #include "EnumParser.h"
 #include "ValueRefParser.h"
+#include "EnumValueRefRules.h"
 #include "EffectParser.h"
 #include "ConditionParserImpl.h"
 #include "MovableEnvelope.h"
